+++ internal/xsql/badconn/badconn_go1.18.go
@@ -22,7 +22,8 @@ func (e Error) Error() string {
 }
 
 func (e Error) Is(err error) bool {
-	if err == driver.ErrBadConn {
+	//nolint:nolintlint
+	if err == driver.ErrBadConn { //nolint:errorlint
 		return true
 	}
 	return xerrors.Is(e.err, err)