+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 * Added retry policy options for topics: `topic/topicoptions.WithReaderCheckRetryErrorFunction`, `topic/topicoptions.WithReaderStartTimeout`, `topic/topicoptions.WithWriterCheckRetryErrorFunction`, `topic/topicoptions.WithWriterStartTimeout`
+* Refactored `internal/conn` middlewares for `DRY`
+* Added `trace.tableSessionInfo.LastUsage()` method for get last usage timestamp
 
 ## v3.41.0
 * Added option for set interval of auth token update in topic streams